Initial commit of v3 branch
This commit adds the AGPLv3 text, an autotools skeleton and a basic
folder structure.

Why rewrite JiyuuBot _again_? The migration to libpurple brought quite a
few unexpected problems, generated primarily by libpurple's poor
documentation. The intention was to create a codebase that was totally
protocol-blind, but what resulted was one that still contained
IRC-specific code. I had attempted a cleanup, starting with enabling
valac's --experimental-non-null flag, but the dearth of libpurple
documentation means it's not trivial to judge what functions may or may
not return null.

Given this, I've had a look around and the Telepathy framework looks
like the way to go.