Add jiyuubot-launcher utility
Since Telepathy is designed for usage by a human being on a desktop,
the session DBus can be 'tainted' by a user's already existing accounts,
installed clients, etc. So this commit adds jiyuubot-launcher (similar
to dbus-launcher) which builds a small DBus environment in /tmp and
spawns a new bus, allowing JiyuuBot and clients a clean environment to
operate in.