Remove telepathy-glib dependency
The telepathy-glib library has a lot of awesome functionality, but
unfortunately it has a rather rigid opinion of how a user is allowed to
interact with Telepathy DBus services. I was getting bogged down in
fighting with it, so I decided to use Vala's awesome DBus support to
write a thin wrapper around the DBus API and ditch telepathy-glib
altogether.