Implement client shutdown timeout
This commit adds some extra logic to libjiyuubot's ClientApp class: time
between messages is measured and checked every minute or so. If the time
elapsed is greater than a pre-defined value (currently ten minutes), the
client quits. This is to prevent many client processes eating resources
listening to an idle bot.

It would be good if this could be configured in the future, but we don't
have the required config infrastructure just yet.