Relax jiyuubot-launcher warning spew
At the moment, if a path is included in jiyuubot-launcher's DBus service
search twice then a 'file exists' warning will be printed for every
service in that directory. This commit checks whether errno is EEXISTS
before issuing a warning.

I had considered using a GenericSet to preclude the possibility of
searching the same path twice, but this won't catch things like
symlinks. This isn't a problem in and of itself, but the volume of
errors displayed contributes a lot of noise and makes it hard to work
out exactly what is happening.